+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 1. [#240](https://github.com/influxdata/influxdb-client-csharp/pull/240): Add IsMeasurement option to Column attribute for dynamic measurement names in POCO classes
 1. [#246](https://github.com/influxdata/influxdb-client-csharp/pull/246), [#251](https://github.com/influxdata/influxdb-client-csharp/pull/251): Add support for deserialization of POCO column property types with a "Parse" method, such as Guid
 
+### Dependencies
+1. [#252](https://github.com/influxdata/influxdb-client-csharp/pull/252): Update dependencies:
+   - NUnit to 3.13.2
+   - NUnit3TestAdapter to 4.0.0
+
 ## 3.0.0 [2021-09-17]
 
 ### Breaking Changes